Standardize EntityDescription in DSMR
