--- homeassistant/components/dsmr/models.py
@@ -1,14 +1,0 @@
-"""Models for the DSMR integration."""
-from __future__ import annotations
-
-from dataclasses import dataclass
-
-from homeassistant.components.sensor import SensorEntityDescription
-
-
-@dataclass
-class DSMRSensorEntityDescription(SensorEntityDescription):
-    """Represents an DSMR Sensor."""
-
-    dsmr_versions: set[str] | None = None
-    is_gas: bool = False
